+++ Sources/TMDB/Configuration/ImageSize.swift
@@ -1,0 +1,3 @@
+public protocol ImageSize<ImagePath>: RawRepresentable<String> {
+    associatedtype ImagePath: RawRepresentable<String>
+}
