+++ Sources/TMDB/Configuration/PosterSize.swift
@@ -1,0 +1,9 @@
+public struct PosterSize: Codable, Hashable, ImageSize, Sendable {
+    public typealias ImagePath = PosterPath
+
+    public var rawValue: String
+
+    public init(rawValue: String) {
+        self.rawValue = rawValue
+    }
+}
